Rename file with typo
